Add 'diff' package to dependencies, enhance error handling in text checking, and modularize App component structure. Update package-lock.json and refactor App.js to utilize new components for improved readability and maintainability.
