+++ src/handlers/text.js
@@ -26,11 +26,17 @@ async function checkSelectedText(shouldShowWindow = false) {
       }
     } catch (error) {
       console.error('Error checking OpenAI client:', error);
-      mainWindow.show();
-      mainWindow.webContents.send('show-settings');
-      mainWindow.webContents.send('show-notification', 'Error connecting to OpenAI. Please check your API key.');
+      // Check if window still exists before sending IPC messages
+      if (windowManager.isWindowAvailable()) {
+        mainWindow.show();
+        mainWindow.webContents.send('show-settings');
+        mainWindow.webContents.send('show-notification', 'Error connecting to OpenAI. Please check your API key.');
+      }
       return;
     }
+    
+    // Check window availability before sending messages
+    if (!windowManager.isWindowAvailable()) return;
     
     // Switch to Grammar & Style tab
     mainWindow.webContents.send('set-active-tab', 0); // Index 0 for Grammar & Style tab
@@ -41,11 +47,16 @@ async function checkSelectedText(shouldShowWindow = false) {
     // Always notify about the text status
     if (!selection) {
       console.log('No text in clipboard');
-      mainWindow.webContents.send('show-notification', 'Please select some text first');
+      // Check window availability again before sending notification
+      if (windowManager.isWindowAvailable()) {
+        mainWindow.webContents.send('show-notification', 'Please select some text first');
+      }
       return;
     }
 
     // Notify about checking start
+    // Check window availability before sending
+    if (!windowManager.isWindowAvailable()) return;
     mainWindow.webContents.send('check-text-reply', { text: selection, corrections: [], isChecking: true });
 
     console.log('Checking text:', selection);
@@ -53,6 +64,9 @@ async function checkSelectedText(shouldShowWindow = false) {
     // Use the OpenAI service for checking
     try {
       const response = await openaiService.checkText(selection);
+      // Check window availability before sending results
+      if (!windowManager.isWindowAvailable()) return;
+      
       if (response.corrections && Array.isArray(response.corrections)) {
         if (response.corrections.length === 0) {
           mainWindow.webContents.send('show-notification', 'Perfect! No corrections needed ✨');
@@ -70,18 +84,26 @@ async function checkSelectedText(shouldShowWindow = false) {
             isChecking: false
           });
         }
+      } else {
+        // Handle cases where response structure might be unexpected but not an error
+        mainWindow.webContents.send('check-text-reply', { text: selection, corrections: [], isChecking: false });
+        mainWindow.webContents.send('show-notification', 'Received an unexpected response format from OpenAI.');
       }
     } catch (error) {
       console.error('Error checking text with OpenAI:', error);
-      mainWindow.webContents.send('check-text-reply', { text: selection, corrections: [], isChecking: false });
-      mainWindow.webContents.send('show-notification', 'Error checking text. Please try again.');
+      // Check window availability before sending error notification
+      if (windowManager.isWindowAvailable()) {
+        mainWindow.webContents.send('check-text-reply', { text: selection, corrections: [], isChecking: false });
+        mainWindow.webContents.send('show-notification', 'Error checking text. Please try again.');
+      }
     }
   } catch (error) {
-    console.error('Error checking text:', error);
-    const mainWindow = windowManager.getMainWindow();
-    if (mainWindow) {
-      mainWindow.webContents.send('check-text-reply', { text: '', corrections: [], isChecking: false });
-      mainWindow.webContents.send('show-notification', 'Error checking text. Please try again.');
+    console.error('Error checking text (outer catch):', error);
+    // Use the isWindowAvailable check here too
+    if (windowManager.isWindowAvailable()) {
+      const currentMainWindow = windowManager.getMainWindow(); // Re-get in case it changed
+      currentMainWindow.webContents.send('check-text-reply', { text: '', corrections: [], isChecking: false });
+      currentMainWindow.webContents.send('show-notification', 'An error occurred while checking text. Please try again.');
     }
   }
 }
